Partially replicated field design

Partially replicated field design

Usage

rand_Prep(
  tot,
  ins,
  rowD,
  trial,
  n_dummies = 0,
  loc = NULL,
  totReps,
  trtrepP,
  trtgroup,
  block_lst,
  season = "season-2024-2025",
  path = t_dir,
  check = c("Shangi", "Unica", "Sagitta", "Sherekea"),
  dummy = c("Unica", "Shangi"),
  to_add = 4
)

Arguments

tot

integer. total number of unique clones/genotypes to be randomized to field

ins

a character vector of genotypes

rowD

integer. number of rows in the field

trial

character. trial name

n_dummies

integer. number of dummies to complete a rectangular layout

loc

character. trial location

totReps

integer. total number of plots: row by col

trtrepP

numeric vector. replications of ins given in the form rep(c(vector of reps), c(vector of number of clones)) e.g., rep(c(1,8), c(304, 4))

trtgroup

numeric vector. replication of treatment groups. samilar repliction as trtrepP but should add to tot.

block_lst

a list specifying blocking of the field

season

season of trial

path

character specifying path to write the design

check

a character vector of checks to fill rectangular grid

dummy

a character vector of dummy checks to fill rectangular grid

to_add

integer. number of checks to add to complete the rectangular grid