resolvable row column design: plots = row by col; test clones replicated once, checks more than once

resolvable row column design: plots = row by col; test clones replicated once, checks more than once

Usage

randomize_res_row_col(
  clones,
  tot,
  trial,
  totReps,
  trtrepP,
  block_lst,
  rowD,
  n_dummies = 0,
  season,
  path = t_dir,
  check = c("Shangi", "Unica", "Sagitta", "Sherekea"),
  dummy = c("Unica", "Shangi"),
  to_add = 4
)

Arguments

clones

a dataframe with geno column

tot

integer. total number of unique clones/genotypes to be randomized to field

trial

character. trial

totReps

integer. total number of plots: row by col

trtrepP

numeric vector. replications of ins given in the form rep(c(vector of reps), c(vector of number of clones)) e.g., rep(c(1,8), c(304, 4))

block_lst

a list specifying blocking of the field

rowD

integer. number of rows in the field

n_dummies

integer. number of dummies to complete a rectangular layout

season

season of trial

path

character specifying path to write the design

check

a character vector of checks to fill rectangular grid

dummy

a character vector of dummy checks to fill rectangular grid

to_add

integer. number of checks to add to complete the rectangular grid

Examples

data("ilri")
dat <- ilri %>% dplyr::select(geno) 
# dat %>% dplyr::pull() 

rrc80 <- randomize_res_row_col(
  clones = dat,
  tot = 62,
  trial = "KE24ILR-BW-ST01",
  totReps = 80,
  trtrepP = rep(c(1, 4, 3), c(55, 4, 3)),
  block_lst = list(c(16,5), c(8,5)),
  rowD = 16,
  # rowsinR = 4,
  # colsinR = 1,
  n_dummies = 3,
  season = "season-2025",
  path = NULL
  
)
#> Warning: replacing previous import 'R.oo::throw' by 'R.methodsS3::throw' when loading 'DiGGer'
#>      Phase,    Search%,    A-measure
#> [1] 1.000000 0.000000 1.321959
#> [1]  1.000000 10.000000  1.303753
#> [1]  1.000000 20.000000  1.303753
#> [1]  1.000000 30.000000  1.303753
#> [1]  1.000000 40.000000  1.303753
#> [1]  1.000000 50.000000  1.303753
#> [1]  1.000000 60.000000  1.303753
#> [1]  1.000000 70.000000  1.303753
#> [1]  1.000000 80.000000  1.303753
#> [1]  1.000000 90.000000  1.303753
#> [1]   1.000000 100.000000   1.303753
#>  [1] 1.303753 1.303753 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9] 0.000000 0.000000
#> 8 rows by 5 columns
#>      Phase,    Search%,    A-measure
#> [1] 2.00000 0.00000 1.99169
#> [1]  2.000000 10.000000  1.633634
#> [1]  2.000000 20.000000  1.632903
#> [1]  2.0000 30.0000  1.6315
#> [1]  2.0000 40.0000  1.6315
#> [1]  2.000000 50.000000  1.630169
#> [1]  2.000000 60.000000  1.630169
#> [1]  2.000000 70.000000  1.630169
#> [1]  2.000000 80.000000  1.630169
#> [1]  2.000000 90.000000  1.628684
#> [1]   2.000000 100.000000   1.628684
#>  [1] 1.628684 1.628684 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9] 0.000000 0.000000

head(rrc80$fieldbook)
#>                    plot          geno entry row column rep
#> 1 KE24ILR-BW-ST01-00001 CIP319027.006    20   1      1   1
#> 2 KE24ILR-BW-ST01-00002 CIP319020.003     7   2      1   1
#> 3 KE24ILR-BW-ST01-00003 CIP319040.004    25   3      1   1
#> 4 KE24ILR-BW-ST01-00004  dummy-Shangi    58   4      1   1
#> 5 KE24ILR-BW-ST01-00005 CIP319064.007    54   5      1   1
#> 6 KE24ILR-BW-ST01-00006 CIP319020.006    10   6      1   1