A wrapper for readr's read_csv & readxl's read_excel for reading trial data

A wrapper for readr's read_csv & readxl's read_excel for reading trial data

read_workbooks(
  dir = t_dir,
  season = "Season 2023",
  read_subfolder = FALSE,
  s = ".csv|.xlsx",
  sub_folder = "Data/Curated Data",
  subset = FALSE,
  n = NULL,
  multiple = FALSE,
  sheet_name = NULL,
  skip = 0
)

Arguments

dir

trial data home directory

season

trial season

read_subfolder

logical. TRUE if data lives inside a subfolder in season

s

a regular expression used to keep relevant files

sub_folder

character of subfolder containing data e.g "Data/Curated Data"

subset

logical. TRUE if you want to read just a subset of files

n

integer. Number of files to subset e.g, 1, -1 1:4. negative prefix if you want to exclude a file at a specific index

multiple

logical. TRUE if you want to read nultiple excel sheets from the same workbook

sheet_name

character. used to specify excel sheet name of interest, e.g., "fieldbook"

skip

integer. number of rows to skip when reading data from an excel workbook

Value

a list of trial dataframes